requirements.txt
PyPDF2==3.0.1
pdfplumber==0.7.5
python-docx==0.8.11
markdown==3.4.4
pandas==1.5.3
